switch to PCG random for p5
Implement random number routines using PCG random

+++ primary/ticktopic.h
@@ -67,7 +67,7 @@ public:
 
     int64_t getInterval() { return fInterval; }
     void setInterval(const uint64_t newInterval) { fInterval = newInterval; }
-    uint64_t setFrequency(const size_t freq) { fInterval = ((uint64_t)(1000.0 / freq)); return fInterval; }
+    uint64_t setFrequency(const float freq) { fInterval = ((uint64_t)(1000.0 / freq)); return fInterval; }
     int64_t getTickCount() { return fTickCount; }
     
     uint64_t getDroppedTicks() { return fDroppedTicks; }
